fix: `setquota` and `delmailuser` (alias removal)
The original `setquota` prompt didn't actually prompt (`-p`), but used `-s` which silents the input, meant for passwords..

Added extra comments to clarify Dovecot quota value retrieved is in units of kibibytes, which is what we multiply by to get the total number of bytes for converting to IEC unit.

---

In `delmailuser`, the shared alias removal code I had it used did not have parity with the full entry removal pattern as a blank alias key was required (due to `^`), corrected by falling back to pattern matching non-whitespace (the alias key) for when the alias arg is empty.