chore: `_create_lock` is not reliable
In it's present form this only locks the command script itself, not a target file (Database) from other commands.

Kept it here as a reminder. It would be more appropriate now in `_db_operation`, however the current locking support would require more intricate lock releasing in that method, and would not be as ideal as locking the files prior to the loop starting..