tests(fix): Correctly use assertions
Assertions often depend on the output of the last `run` command.

These tests were either missing a `run`, or incorrectly asserting against the wrong command they were intended for.

`wait_for_service` should not be asserted. It is just a pre-requisite before the actual test case is ready to proceed.

Updating bats and submodules helped surface this testing bug.