tests(fix): Correctly check `LOG_LEVEL` ENV and `changedetector` status
`/etc/dms-settings` quote wraps the value, but the assert wasn't helpful either with the failure (no match).

Chose to fetch the ENV from container environment with `env` as we don't rely on `start-mailserver.sh` processing for this ENV anyway. Also switched to `assert_regex` for better failure output.

The `changedetector` state conditions were returning the wrong status code (0 is true/ok in bash). An earlier misunderstanding to resolve a failure had me add explicit status codes to return for true/false, as it didn't seem to work correctly with just the `[[ expression ]]` but was due to something else?