tests(fix): Wait on directory to be removed
Workaround that tries not to introduce heavier delays by waiting on a full change event to complete in the previous `email update` if possible.

There is a chance that the account has the folder deleted, but restored from an active change event (for password update, then the account delete).

---

Additionally apply review feedback to add inline maintainer note about the healthcheck test to avoid refactors making the same mistake.