add clamav-stage for better caching
This commit introduces a new stage (`clamav-stage`) and the `--link`
flag for teh `COPY` directive. We now require `DOCKER_BUILDKIT` to be
set to `1`. Moreover, the `Dockerfile` syntax is explictly set in the
`Dockerfile` (required for `--link`).

A bit of refactoring was done in the Dockerfile to keep everything
consistent.