improve GitHub Action CI with re-usable workflows
A new re-usable workflow was introduced to handle building the container
image in a generic way. Now, the `scheduled_builds`, `default_on_push`
and a part of the `test_merge_requests` workflow can use the same code
for building and publishing the container images. This is DRY.

The `test_merge_requests` workflow was adjusted in the following way:
building the ARM image is now done _after_ the initial build to run in
parallel with BATS. This saves around 13 minutes of CI time. The new ARM
build uses the generic build workflow. The other two jobs do not use the
generic workflow yet as I was not sure whether the generic workflow fits
our needs in this case.