add script that install packages
This new script is a clean way of handling the installation of packages.
I think the huge `RUN` command in `Dockerfile` was hard to read and
maintain.

Using a script is a non-issue, as the image is rebuilt whenever the
script is touched.

Moreover, this commit implicitly removed the explicit Fail2Ban
installation and therefore #1965, BECAUSE Debian 11 now ships with the
latest and greatest Fail2Ban version (even more up-to-date than out
current version it seems).