update ´VERSION` and `CHANGELOG.md`
I introduced the "Keep a Changelog"-format which helps me in **not**
spending this much time on changelogs in the future. There is a another
PR that will update the documentation before v11.2.0 is released, which
focuses on adding the entry to the changelog before merging a relevant
PR.

Releasing is always a huge efforts when I get the changelog and adjus
it. This is incceptable and needs to change. This is the solution.