update contributing documentation
This PR updates the contribution documentation. It adds the information
to add your changes to the changelog yourself (see the v11.2.0 release
PR (which may be coming up in the future, depending on when you read
this)). Moreover, it refactores teh tests, documentation and coding
style pages. These were bloated or outdated, and have been merged into
one site. This seems to be the clean solution to me.