properly provide ENV
last commit was done in error; now the `IMAGE_NAME` should be provided
properly to the tests. `IMAGE_NAME` should be the new ENV variable name
for new tests because `NAME` is a bit too generic. I'm just introducing
it here, we can change it step by step in the future.