update to latest version of Fail2Ban
Tests pass locally; we just need to add a "fix". The new Fail2Ban
configuration complains about `allowipv6` not being set in the default
configuration. Hence, I set it in a custom configuration. There is an
additional description of this issue in the file itself. Other than
that, all looks good.