run `source` instead of `/bin/bash` for user patches
Running `source` will "execute" the file as well, but it allows
tinkering with existing code. I have a use case where I want to override
a specific function, which I can do with `source`, but I cannot right
now, because it is run in a "subshell", where changes are not
propagates.

This does NOT change the current behavior, it just opens up more
possibilities.