apply PR feedback
I removed the long options from the command and replaced them with their
short versions. This allowed me to then nicely add line breaks to
increase readability of the command.

I couldn't withstand reworking the above command too; it is now using a
proper array as well (with proper splitting before).

Moreover, I added the `dms-test_` prefix to the linting tests so we do
not need to grep for them anymore, which shortens the commands.

If you have 100 characters in width for the `Makefile`, the clean target
should be properly displayed now.